# openapi/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/privacy-policies": {
      "get": {
        "operationId": "PublicPrivacyPoliciesController_renderPoliciesList",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PublicPrivacyPolicies"
        ]
      },
      "post": {
        "operationId": "PrivacyPoliciesController_createPolicy",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePrivacyPolicyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Privacy policy created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyPolicyDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "409": {
            "description": "Privacy policy with this ID already exists"
          }
        },
        "summary": "Create a new privacy policy",
        "tags": [
          "privacy-policies"
        ]
      }
    },
...
    "/api/privacy-policies/{id}": {
      "get": {
        "operationId": "PrivacyPoliciesController_getPolicyById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Privacy policy ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Privacy policy found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyPolicyDto"
                }
              }
            }
          },
          "404": {
            "description": "Privacy policy not found"
          }
        },
        "summary": "Get privacy policy by ID",
        "tags": [
          "privacy-policies"
        ]
      },
      "put": {
        "operationId": "PrivacyPoliciesController_updatePolicy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Privacy policy ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePrivacyPolicyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Privacy policy updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyPolicyDto"
                }
              }
            }
          },
          "404": {
            "description": "Privacy policy not found"
          }
        },
        "summary": "Update an existing privacy policy",
        "tags": [
          "privacy-policies"
        ]
      },
      "delete": {
        "operationId": "PrivacyPoliciesController_deletePolicy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Privacy policy ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Privacy policy deleted successfully"
          },
          "404": {
            "description": "Privacy policy not found"
          }
        },
        "summary": "Delete a privacy policy",
        "tags": [
          "privacy-policies"
        ]
      }
    },
    "/api/privacy-policies/app/{appId}": {
      "get": {
        "operationId": "PrivacyPoliciesController_getPolicyByAppId",
        "parameters": [
          {
            "name": "appId",
            "required": true,
            "in": "path",
            "description": "Application ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Privacy policy found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyPolicyDto"
                }
              }
            }
          },
          "404": {
            "description": "Privacy policy not found"
          }
        },
        "summary": "Get privacy policy by app ID",
        "tags": [
          "privacy-policies"
        ]
      }
    },
    "/api/privacy-policies/app/name/{appName}": {
      "get": {
        "operationId": "PrivacyPoliciesController_getPolicyByAppName",
        "parameters": [
          {
            "name": "appName",
            "required": true,
            "in": "path",
            "description": "Application name",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Privacy policy found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyPolicyDto"
                }
              }
            }
          },
          "404": {
            "description": "Privacy policy not found"
          }
        },
        "summary": "Get privacy policy by app name",
        "tags": [
          "privacy-policies"
        ]
      }
    },
    "/api/privacy-policies/{id}/content": {
      "get": {
        "operationId": "PrivacyPoliciesController_getPolicyContent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Privacy policy ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Privacy policy content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "content": {
                      "type": "string"
                    },
                    "lastUpdated": {
                      "type": "string"
                    },
                    "version": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Privacy policy not found"
          }
        },
        "summary": "Get privacy policy content only",
        "tags": [
          "privacy-policies"
        ]
      }
    },
    "/api/privacy-policies/search/query": {
      "get": {
        "operationId": "PrivacyPoliciesController_searchPolicies",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "required": false,
            "in": "query",
            "description": "Language filter",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PrivacyPolicyDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Search privacy policies",
        "tags": [
          "privacy-policies"
        ]
      }
    },
    "/api/privacy-policies/web": {
      "get": {
        "operationId": "PrivacyPoliciesController_renderPoliciesList",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Render privacy policies list page",
        "tags": [
          "privacy-policies"
        ]
      }
    },
    "/api/privacy-policies/web/{id}": {
      "get": {
        "operationId": "PrivacyPoliciesController_renderPolicyPage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Privacy policy ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Render privacy policy page",
        "tags": [
          "privacy-policies"
        ]
      }
    },
    "/api/privacy-policies/web/app/{appId}": {
      "get": {
        "operationId": "PrivacyPoliciesController_renderPolicyPageByAppId",
        "parameters": [
          {
            "name": "appId",
            "required": true,
            "in": "path",
            "description": "Application ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Render privacy policy page by app ID",
        "tags": [
          "privacy-policies"
        ]
      }
    },
    "/api/privacy-policies/web/app/name/{appName}": {
      "get": {
        "operationId": "PrivacyPoliciesController_renderPolicyPageByAppName",
        "parameters": [
          {
            "name": "appName",
            "required": true,
            "in": "path",
            "description": "Application name",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Render privacy policy page by app name",
        "tags": [
          "privacy-policies"
        ]
      }
    },
...
  },
  "info": {
    "title": "Doctopus API",
    "description": "Modern NestJS API with comprehensive features",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    },
    {
      "url": "https://tree2u.kobotogether.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "PrivacyPolicyDto": {
        "type": "object",
        "properties": {}
      },
      "CreatePrivacyPolicyDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePrivacyPolicyDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}